server/requirements.txt
mcp[cli]>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.32.0
fastapi>=0.115.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
watchfiles>=0.21.0
croniter>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
cryptography>=41.0.0
-e .
# Phase 6: Interface Expansion
openai>=1.0.0
rich>=13.0.0
